bmkgsatu-sender/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
loguru==0.7.2
pika==1.3.2
python-decouple==3.8
requests==2.32.3
tenacity==9.0.0
urllib3==2.2.3
